+++ inc/theme-settings.php
@@ -15,14 +15,23 @@ namespace Secretum;
 
 /**
  * WordPress Required Content Width
+ * Set the content width in pixels, based on the theme's design and stylesheet.
+ * Priority 0 to make it available to lower priority callbacks.
  *
- * @since 1.0.0
+ * @global int $content_width Content width.
  *
  * @link https://codex.wordpress.org/Content_Width
+ *
+ * @since 1.0.0
  */
-if ( true !== isset( $content_width ) ) {
-	$content_width = secretum_mod( 'secretum_content_width' ) ? absint( secretum_mod( 'secretum_content_width', 'int' ) ) : 640;
+function secretum_content_width() {
+	// This variable is intended to be overruled from themes.
+	// Open WPCS issue: {@link https://github.com/WordPress-Coding-Standards/WordPress-Coding-Standards/issues/1043}.
+	// phpcs:ignore WordPress.NamingConventions.PrefixAllGlobals.NonPrefixedVariableFound
+	$GLOBALS['content_width'] = secretum_mod( 'secretum_content_width' ) ? absint( secretum_mod( 'secretum_content_width', 'int' ) ) : 640;
 }
+
+add_action( 'after_setup_theme', 'Secretum\secretum_content_width', 0 );
 
 
 /**